Rename PrimOpsList.h to prim-ops-list.h
---
